Extend strict mypy checks
Strict mypy checks cannot be enabled on a per-module level, so we used
the disallow_untyped_defs option to replace them.  But actually there
are many more options that would be set in strict mode.  This patch adds
these options to the mypy configuration, fixes some missing or wrong
annotations and adds some ignores for more complex issues so that the
checks still pass.

Fixes: https://github.com/Nitrokey/pynitrokey/issues/443